% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_extravignette.R
\name{build_extravignette}
\alias{build_extravignette}
\title{Build optional vignette once the package is installed}
\usage{
build_extravignette()
}
\description{
This function aims at creating vignettes after the package installation.
}
\examples{
search()
build_extravignette()

}
